requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.txt
#
attrs==20.3.0             # via -r requirements-dev.txt, pytest
cuelib==0.2.0             # via -r requirements-dev.txt
iniconfig==1.1.1          # via -r requirements-dev.txt, pytest
mypy-extensions==0.4.3    # via -r requirements-dev.txt, mypy
mypy==0.790               # via -r requirements-dev.txt
packaging==20.4           # via -r requirements-dev.txt, pytest
pexpect==4.8.0            # via -r requirements-dev.txt
pluggy==0.13.1            # via -r requirements-dev.txt, pytest
ptyprocess==0.6.0         # via -r requirements-dev.txt, pexpect
py==1.9.0                 # via -r requirements-dev.txt, pytest
pyparsing==2.4.7          # via -r requirements-dev.txt, packaging
pytest-asyncio==0.14.0    # via -r requirements-dev.txt
pytest==6.1.2             # via -r requirements-dev.txt, pytest-asyncio
six==1.15.0               # via -r requirements-dev.txt, packaging
toml==0.10.2              # via -r requirements-dev.txt, pytest
typed-ast==1.4.1          # via -r requirements-dev.txt, mypy
typing-extensions==3.7.4.3  # via -r requirements-dev.txt, mypy
